# internal/server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gemba HTTP API",
    "version": "0.1.0",
    "description": "Wire schema for the gemba server's `/api/*` surface (gm-e4.2). The SPA at `web/` consumes a typed client generated from this document; the source of truth lives in `docs/api/openapi.json` and is served at runtime under `/api/openapi.json`.\n\nThis MVP covers the read endpoints the SPA depends on day-to-day: health, version, config, repositories, work items, sprints, agents, sessions, escalations, capabilities, and the new-project conversational flow. The remaining `/api/*` surface (consults, walks, planner, perspectives) is acknowledged in `paths` with summary-only stubs and will gain full request/response schemas in follow-up beads.",
    "contact": {
      "name": "GembaCore/gemba-core",
      "url": "https://github.com/GembaCore/gemba-core"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin (the gemba server is the SPA host)."
    }
  ],
...
  "paths": {
...
    "/api/work-items/{id}": {
      "get": {
        "tags": ["work"],
        "summary": "Fetch one WorkItem by id.",
        "operationId": "getWorkItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Workspace-prefixed bead id (e.g. `gemba/gm-e4.2`).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The work item.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorkItem" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/AdaptorUnavailable" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "AdaptorUnavailable": {
        "description": "Adaptor is not registered or is degraded; the SPA shows a banner and keeps cached data visible.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Shared error envelope (gm-faz). `error` is a stable token the SPA branches on; `message` is human-readable and may be surfaced verbatim.",
        "required": ["error", "message"],
        "properties": {
          "error": { "type": "string", "examples": ["adaptor_degraded", "validation", "not_found"] },
          "message": { "type": "string" }
        }
      },
...
      "StateCategory": {
        "type": "string",
        "enum": ["backlog", "unstarted", "staged", "started", "completed", "canceled"]
      },
      "AgentRef": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" }
        }
      },
      "WorkItem": {
        "type": "object",
        "description": "Subset of the WorkItem fields most commonly read by the SPA. The full shape (relationships, evidence, dod, derived signals, custom fields) is the source of truth in `core/types.go` and `web/src/types/core.gen.ts`; this schema documents the wire-stable subset for a typed client. `additionalProperties: true` keeps clients forward-compatible with fields landing in follow-up beads.",
        "additionalProperties": true,
        "required": ["id", "kind", "title", "status", "state_category", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string" },
          "kind": { "type": "string", "examples": ["epic", "task", "feature", "decision"] },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "state_category": { "$ref": "#/components/schemas/StateCategory" },
          "priority": { "type": ["integer", "null"], "minimum": 0, "maximum": 4 },
          "owner": { "$ref": "#/components/schemas/AgentRef" },
          "assignee": { "$ref": "#/components/schemas/AgentRef" },
          "labels": { "type": "array", "items": { "type": "string" } },
          "primary_repository_id": { "type": "string" },
          "repository_ids": { "type": "array", "items": { "type": "string" } },
          "sprint_id": { "type": ["string", "null"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token auth (gm-e5.1). Required when the server boots with `--auth=token`."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gemba_session",
        "description": "Signed session cookie issued by `POST /api/auth/login` after a bearer exchange."
      }
    }
  },
  "security": [
    { "bearerToken": [] },
    { "sessionCookie": [] },
    {}
  ]
}